Crude v1.
Accepts a UPRN on the URL, downloads the PDF and parses the dates from the PDF into a dates array.
This produces a dynamic `VCALENDAR` file which contains combined entries for when collections are on the same day.
Alerts are added to the feed, although it looks like Google Calendar ignores these for now.